LayoutTests/fast/text-autosizing/ios/idempotentmode/idempotent-percentage-expected.txt

PASS result is 12
PASS result is 6
PASS result is 6
PASS result is 12
PASS successfullyParsed is true

TEST COMPLETE
Test
Test
Test
Test
Test
Test
Test
Test
